+++ wham-2.0/src/wham_types.h
@@ -6,6 +6,8 @@
 #include <boost/lexical_cast.hpp>
 #include <boost/numeric/conversion/cast.hpp>
 #include <limits>
+#include <ctime>
+#include <cstdlib>
 
 #ifndef KB
 #define KB 0.00813144621 // kJ/(mol*K)
@@ -17,13 +19,13 @@ struct t_options
     std::string outname;
     int f0;
     int fN;
-    double t;
-    std::vector<double> x0;
-    std::vector<double> xN;
-    std::vector<double> b;
+    float t;
+    std::vector<float> x0;
+    std::vector<float> xN;
+    std::vector<float> b;
     int ndof;
-    double tol;
-    double iter;
+    float tol;
+    float iter;
     int ntraj;
     bool bVerbose;
     bool doConv;
@@ -46,19 +48,19 @@ struct t_wham
     int nstates;
     std::vector<int> sample;
     std::vector<int> counts;
-    std::vector<double> t;
+    std::vector<float> t;
     std::vector<std::vector<double> > omegas;
 };
 
 struct t_experiment
 {
     int experiment;
-    double t;
+    float t;
     // [dofM]
     std::vector<int> nbins;
-    std::vector<double> phi0, deltaphi, fc;
+    std::vector<float> phi0, deltaphi, fc;
     // [frameN][dofM]
-    std::vector<std::vector<double> > potential, coordinate;
+    std::vector<std::vector<float> > potential, coordinate;
     std::vector<std::vector<t_bin> > frame_bin;
     // [binL][dofM]
     std::vector<std::vector<double> > omega;
